backend/pyproject.toml
[project]
name = "repobuddy-backend"
version = "0.1.0"
description = "RepoBuddy — Intelligent Codebase Understanding Platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "pgvector>=0.3.5",
    "redis>=5.1.0",
    "celery[redis]>=5.4.0",
    "networkx>=3.4",
    "tree-sitter>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.12",
    "orjson>=3.10.0",
    "gitpython>=3.1.43",
    "openai>=1.50.0",
    "tiktoken>=0.7.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
# E501: formatter handles line length; don't double-enforce.
# B008: FastAPI uses Depends()/File()/Query() in default args by design.
# UP042: str enums are still widely used; not worth bulk-converting.
# RUF002/RUF003: unicode chars in docstrings/comments are intentional.
ignore = ["E501", "B008", "UP042", "RUF002", "RUF003", "SIM102"]

[tool.ruff.lint.per-file-ignores]
"alembic/versions/*.py" = ["N999", "E402"]
"tests/*.py" = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
